V4 검수 반영: 안 올림·로그인 화면 통합·오류 문구·가격 스트립·폼 축소
- SKIPPED 표시명 보관 → 안 올림(✔️ 안 올릴게요). 지켜보기·30일 보관과 뜻이 겹치던 것.
- 문서 요청은 401 HTML 대신 /login?next=로 보낸다(같은 출처 경로만). RSC·fetch·서버 액션은 401 그대로.
- 패스키 오류는 현표용 한 줄 + 관리자용 한 줄. 로그인 실패의 첫 후보는 "이 폰을 아직 등록하지 않았다면".
- deal-flow·curator-link·url-guard·profile·watch·renderer 경고 문구 해요체, 기술어 제거.
- 가격 스트립: 평소보다 N% 싸요 / 지금이 N% 더 싸요, 첫 기록만 있으면 그리지 않음.
- 직접 만들기 폼에서 색상별 링크·AI 체크박스 제거. WatchList.tsx(미사용) 삭제.
- 허브·만료 안내의 "살 수 있는 딜" 조건을 lib/hub-query.ts 하나로.

+++ src/lib/format.ts
@@ -17,7 +17,7 @@ const KST_ENDS_AT = new Intl.DateTimeFormat("ko-KR", {
 
 const KST_SHORT = new Intl.DateTimeFormat("ko-KR", {
   timeZone: "Asia/Seoul",
-  month: "numeric",
+  month: "long", // "9월 18일 14:49" — 목록의 날짜 그룹 제목("9월 16일 (수)")과 같은 꼴
   day: "numeric",
   hour: "2-digit",
   minute: "2-digit",
